% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CreateNetwork.R
\name{CreateNetwork}
\alias{CreateNetwork}
\title{Laplacian Matrix for Prior Biological Knowledge in Network Constraint}
\usage{
CreateNetwork(
  X,
  Y = NULL,
  delta = NULL,
  doid = NULL,
  tissue = NULL,
  disease_file = NULL,
  tissue_file = NULL,
  cache = FALSE,
  cache_dir = NULL,
  choice = 1,
  model = NULL,
  dist = NULL,
  verbose = FALSE
)
}
\arguments{
\item{X}{Numeric matrix of standardized covariates.}

\item{Y}{Numeric vector of observed survival times (log-transformed under \code{AFTNet}),
required for \code{choice = 2}.}

\item{delta}{Integer vector of censoring indicators (1 = event, 0 = censored),
required for \code{choice = 2}.}

\item{doid}{Character string specifying Disease Ontology ID (\code{"DOID:XXXX"}),
used only if \code{disease_file} is not provided.}

\item{tissue}{Character string specifying tissue name, used to retrieve the
tissue-specific network from HumanBase, used only if
\code{tissue_file} is not provided.}

\item{disease_file}{Character string specifying optional path to a tab-delimited
file containing disease-associated genes (columns: \code{entrez_id},
\code{standard_name}, and \code{score}).}

\item{tissue_file}{Character string specifying optional path to a tab-delimited
file with tissue-specific gene interactions (columns:
\code{gene1}, \code{gene2}, and \code{score}).}

\item{cache}{Logical value; if \code{TRUE}, downloaded HumanBase files are cached
for reuse in \code{cache_dir}. If \code{FALSE} (default), files are downloaded
for the current session only.}

\item{cache_dir}{Character string specifying a directory used to cache
downloaded HumanBase files (when \code{cache = TRUE}).}

\item{choice}{Value specifying the choice for the signs of the adjacency matrix
\itemize{
\item \code{1} (default): for correlation-based signs.
\item \code{2}: for ridge-based signs.
}}

\item{model}{Character string specifying the fitted survival model
(\code{"COXNet"}, or \code{"AFTNet"}) required only for \code{choice = 2}.}

\item{dist}{Character string specifying the AFTNet distribution.
Must be one of \code{"weibull"}, \code{"lognormal"}, or \code{"loglogistic"}
(required only for \code{choice = 2}).}

\item{verbose}{Logical value, if \code{TRUE} progress messages are printed.}
}
\value{
A list with two elements:
\itemize{
\item \code{disease_genes}: data frame of disease genes used in the network.
\item \code{L}: final Laplacian matrix.
}
}
\description{
Builds a Laplacian network penalty based on a prior weighted graph.
It encourages coefficients corresponding to connected covariates to behave similarly:
if two covariates are strongly connected in the network, their estimated coefficients
tend to be either both close to zero or both nonzero. In this way, the penalty promotes
smoothness and structural coherence across related variables.
}
\details{
This prior network is represented by a weighted graph where each vertex
corresponds to a covariate and the edges describe relationships between covariates.
The edge weights are stored in an adjacency matrix \eqn{A}, which has zeros on its diagonal.
The degree matrix \eqn{D} contains on its diagonal the sum of the absolute
edge weights connected to each vertex. The Laplacian matrix is defined as \eqn{L = D - W},
where \eqn{W} is the weighted matrix estimated from \eqn{A}.
Two strategies can be used.
\itemize{
\item Correlation-based signs (\code{choice = 1}): the sign of an edge is
set according to the Pearson correlation between the two corresponding covariates.
\item Ridge-based signs (\code{choice = 2}): the sign of an edge is determined
by the signs of ridge regression coefficients obtained from a penalized
survival model. This ridge estimator provides stable coefficient estimates
in high-dimensional settings. For the Cox model the ridge fit is obtained
via \code{glmnet::glmnet()}), while for the AFT model via \code{survival::survreg()}).
}

The framework is used to construct a disease-specific gene interaction network,
where edges represent biological relationships between genes relevant to a
given cancer and tissue type.

Internally, the function relies on helper routines (see \code{\link{RepositoryDisease}} and \code{\link{RepositoryTissue}})
to retrieve biological prior information from the \href{https://hb.flatironinstitute.org/}{HumanBase} database.
These datasets are combined to construct a disease- and tissue-specific adjacency matrix
that defines the structure of the Laplacian penalty. User-provided files with
the same format can be supplied to bypass the download step.
}
\note{
If tissue-specific or disease-specific files are not provided, the function
downloads the relevant data from HumanBase. In this case, an active internet
connection is required. Moreover, not all DOIDs and tissues are present in the
HumanBase repository. f the requested is not available, the function may return
an empty list.
}
\examples{
  \donttest{
  
    data(LUADdataset)
  
    net <- CreateNetwork(
              LUADdataset$X_train,
              doid    = "DOID:1324",
              tissue  = "lung",
              choice  = 1,
              verbose = TRUE)
              
    L   <- net$L                          # final laplacian matrix
  
    disease_genes <- net$disease_genes    # disease genes and scores
  
  }
  
}
