% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CreateNetwork.R
\name{RepositoryTissue}
\alias{RepositoryTissue}
\title{Tissue-Specific Top Edge Network from HumanBase}
\usage{
RepositoryTissue(
  tissue = NULL,
  cache = FALSE,
  cache_dir = NULL,
  verbose = FALSE
)
}
\arguments{
\item{tissue}{Character string specifying the name of the tissue to download.
Spaces will automatically be converted to underscores.}

\item{cache}{Logical value; if \code{TRUE}, downloaded HumanBase files are cached
for reuse in \code{cache_dir}. If \code{FALSE} (default), files are downloaded
for the current session only.}

\item{cache_dir}{Character string specifying a directory used to cache
downloaded HumanBase files (when \code{cache = TRUE}).}

\item{verbose}{Logical value, if \code{TRUE} progress messages are printed.}
}
\value{
A data.frame with tissue-specific gene interactions (columns:
\code{gene1}, \code{gene2}, and \code{score}).
}
\description{
Downloads the top edge gene interaction network for a specific human tissue
from the \href{https://hb.flatironinstitute.org/download/tissue-networks/giant}{HumanBase} resource.
}
\note{
An active internet connection is required.
}
\examples{
\donttest{

   # - Download tissue-specific repository for Lung Adenocarcinoma -

      tissue <- RepositoryTissue(
       tissue    = "lung",
       cache     = FALSE,
       cache_dir = NULL,
       verbose   = FALSE
      )

      head(tissue)
}

}
